adding display of widgets with config to screen after selecting and naming.
starting to scratch the surface of the modularity aspects of the code.
soon will be at the point where most of the work is in the widgets and
structuring the widgets so they can be effectively managed and easily
created. Pretty sure there are some things wrong with the code as it
stands now, but too tired to think clearly about it.
What we want is for new widget designers to have a clear set of
guidelines and boundaries for the creation of new widgets. the basics
(saving, retrieving, orienting on page) should be handled for them by
defaults, but allow for easy override.
Next up:
- editing the widget config on the thingType page,
— happens in individual widget code
- saving that config to db
— happens in generic code
- loading that config up again from db.
— happens in generic code.