have got thingType defaultWidget Definitions saving and retrieving from DB
Probably some refactoring needed now.
There’s something bad about the way code is divided between the
entities  (Things, widgets), and the db code.

and the way the db code roundtrips to the server when it’s on the
client is a bit bad too.

After refactoring, the next thing is to move on to the thing instance
page itself, where things really get interesting. :)