changed url loader widget to more narrowly defined jsonloader, for a more focused first widget test case
for the record, I’m really not happy with the way the rendering code is
put together; I’d rather have better data binding, templates, etc. But
i could spend weeks just getting that right, and it’s more important to
move on to getting some widgets that work, talk to each other, and save
and load correctly. So we can make pages that look interesting.

1. get some useful widgets done.
2. refactor the widget code so new widgets are easier to create.
3. refactor the other code (configuration pages especially)