added a template renderer widget
This new widget let’s you customize a template using coeditor. The
template can use dot-style template language.

When the template is parsed, all widgets the template widget depends on
will be loaded into the template data, so values in, for example, a
jsonloaded widget, can be rendered in the template. This works for
image urls too.

These two templates can do about 80% of what’s needed, if cleverly
combined.

Widgets I need to make soon:
indexer - to extract relevant data and get it into search indexers,
like solr
comments - for allowing comments on a page
linker - to allow people to add links on any page, of other relevant
pages.
image annotator - for annotating images on the page
widgets that tie in to other mediaLab projects and code.